+++ KitCli.Abstractions/Io/CliIo.cs
@@ -2,18 +2,36 @@ namespace KitCli.Abstractions.Io;
 
 public class CliIo : ICliIo
 {
-    public string? Ask()
-        => Console.ReadLine();
+    public async Task<string?> AskAsync(CancellationToken cancellationToken)
+    {
+        var abandonableBackgroundConsoleRead = Task.Run(Console.ReadLine, CancellationToken.None);
+
+        try
+        {
+            return await abandonableBackgroundConsoleRead.WaitAsync(cancellationToken);
+        }
+        catch (OperationCanceledException)
+        {
+            return null;
+        }
+    }
 
     public void Pause()
         => Console.WriteLine();
-    
+
     public void Say(string something)
         => Console.WriteLine(something);
-    
+
     public void SetTitle(string title)
         => Console.Title = title;
 
-    public void OnCancel(Action cancel) 
-        => Console.CancelKeyPress += (_, _) => cancel();
+    public void OnCancel(Action cancel)
+        => Console.CancelKeyPress += (_, e) =>
+        {
+            SuppressDefaultAbruptTermination(e);
+            cancel();
+        };
+
+    private static void SuppressDefaultAbruptTermination(ConsoleCancelEventArgs e)
+        => e.Cancel = true;
 }