+++ KitCli.Abstractions/Io/ICliIo.cs
@@ -1,8 +1,8 @@
 namespace KitCli.Abstractions.Io;
 
-public interface ICliIo 
+public interface ICliIo
 {
-    string? Ask();
+    Task<string?> AskAsync(CancellationToken cancellationToken);
     void Pause();
     void Say(string something);
     void SetTitle(string title);