+++ KitCli.Workflow.Abstractions/ICliWorkflow.cs
@@ -3,10 +3,14 @@ namespace KitCli.Workflow.Abstractions;
 public interface ICliWorkflow
 {
     CliWorkflowStatus Status { get; }
-    
+
+    CancellationToken CancellationToken { get; }
+
     List<ICliWorkflowRun> Runs { get; }
-    
+
     ICliWorkflowRun NextRun();
 
     void Stop();
+
+    void InterruptCurrentRun();
 }